# bigint: simplify the add2/sub2 loops