@cuviper 2016-06-30
Benchmarks are pretty flat for small numbers, but the larger ones show great improvement, e.g. from:

```
test divide_2          ... bench:   1,285,211 ns/iter (+/- 36,653)
test fib_10000         ... bench:   1,525,243 ns/iter (+/- 90,944)
test multiply_2        ... bench:   3,736,716 ns/iter (+/- 2,741,946)
```

to:

```
test divide_2          ... bench:   1,023,933 ns/iter (+/- 14,726)
test fib_10000         ... bench:     950,906 ns/iter (+/- 31,328)
test fib_10000         ... bench:   1,525,243 ns/iter (+/- 90,944)
```

The motivating examples from #200 look much better too:

Name | Python | Rust before | Rust after
--- | --- | --- | ---
fib (add+sub) | 24.41s | 39.57s | 28.29s
fib2 (add+swap) | 10.32s | 14.62s | 11.67s