@homu 2016-06-30
:pushpin: Commit 609629d has been approved by `hauleth`

<!-- @homu r=hauleth 609629d34d00fde00d22411c7db08e66c106ff0b -->