@cuviper 2016-07-01
It seems homu.io is not fully working lately.  There's talk of replacing it at rust-community/team#41.